Display customers

+++ src/utils/api.js
@@ -1,8 +1,10 @@
 import {create} from 'axios'
+import store from '../store'
+
 
 const api = function (headers = null) {
     let options = {
-        baseURL: process.env.API,
+        baseURL: process.env.VUE_APP_API_URL,
         headers: []
     };
     options['headers'] = headers === null ? {
@@ -12,11 +14,12 @@ const api = function (headers = null) {
     requester.interceptors.response.use((response) => {
         return response;
     }, (e) => {
-        if (e.request.status === 401) {
+        if (e.request.status === 401 && store.getters['user/isAuthenticated']) {
             window.localStorage.clear();
             location.reload();
         }
-        return e.request;
+        //@todo error handler
+        return Promise.reject(e)
     });
     return requester;
 };
